+++ src/main.c
@@ -6,6 +6,17 @@
 #include <readline/history.h>
 #include <unistd.h>
 
+void	msh_exit(t_shell *shell, int exitcode)
+{
+	if (shell->line)
+		free_null(shell->line);
+	ft_printf("exiting now...\n");
+	if (shell->owned_envp)
+		arr_free(shell->owned_envp);
+	free_null(shell);
+	exit(exitcode);
+}
+
 #include "../src/signals/msh_signals.h"
 #include "../src/builtins/builtins.h"
 #include "../src/parser/parser.h"
