+++ test/support/support_parser.h
@@ -12,10 +12,10 @@
 # include "print_arr_sep.c"
 # include "occurs.c"
 
-#include "lexer.c"
-#include "../src/lexer/lexer_support.c"
-#include "../src/lexer/check_quotes.c"
-#include "../src/lexer/check_pipes.c"
-#include "../src/utils/while_string_wrapper.c"
-#include "../src/utils/bool_array.c"
+# include "lexer.c"
+# include "../src/lexer/lexer_support.c"
+# include "../src/lexer/check_quotes.c"
+# include "../src/lexer/check_pipes.c"
+# include "../src/utils/while_string_wrapper.c"
+# include "../src/utils/bool_array.c"
 #endif
